# api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  title: API Survey
  description: API description in Markdown
  version: 1.0.0
host: localhost:3000
...
schemes:
  - http
  - https
parameters: 
  survey:
    in: query
    name: survey
    required: true
    type: integer
    description: survey.
  surveyName:
    in: query
    name: surveyName
    required: true
    type: string
    description: Survey Name
  surveySearch:
    in: path
    name: surveyId
    required: true
    type: integer
    description: survey id
    
  # ANSWER
...
paths:
...
  /{surveyId}:
    get:
      summary: Returns a survey
      tags:
        - Survey
      description: extended description
      operationId: getSurvey
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: surveyId
          required: true
          type: integer
          minimum: 0
      responses:
        '200':
          description: OK
          schema:
             $ref: '#/definitions/SurveyResponseSuccess'
        '400':
          description: Bad request. surveyId must be an integer and larger than 1
          schema:
            $ref: '#/definitions/SurveyResponseFail'
        '500':
          description: Unexpected error.  
          schema:
            $ref: '#/definitions/SurveyResponseError'
...
  /{surveyId}/:
    put:
      summary: Replace a survey
      tags:
        - Survey
      description: extended description
      operationId: updateSurvey
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/surveySearch'
        - $ref: '#/parameters/survey'
        - $ref: '#/parameters/surveyName'
        
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              response:
                type: object
        '400':
          description: "the parametrs should not be empty"
          examples:
            aplication/json: {
              "message": "the parametrs should not be empty",
              "reqId": "2O3n-ffC_"
            } 
          schema:
            $ref: '#/definitions/SurveyResponseFail'
        
        '500':
          description: Unexpected error.    
          schema:
            $ref: '#/definitions/SurveyResponseError'
  /{surveyId}//:
    delete:
      summary: Returns a survey
      tags:
        - Survey
      description: extended description
      operationId: deleteSurvey
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: surveyId
          required: true
          type: integer
          minimum: 1
      responses:
        '200':
          description: OK
          examples:
            aplication/json: { "response":{
              "survey": "{\"pages\":[{\"name\":\"formulario\",\"elements\":[{\"type\":\"text\",\"name\":\"name\",\"title\":\"Nombre\"},{\"type\":\"text\",\"name\":\"lastName\",\"title\":\"Apellido\"},{\"type\":\"checkbox\",\"name\":\"chooseSex\",\"title\":\"sexo\",\"choices\":[{\"value\":\"item1\",\"text\":\"item3\"},\"item2\",\"item3\"]}]}]}",
              "surveyName": "Encuesta entregas",
              "createdAt": "2019-05-03T20:39:07.605Z",
              "updatedAt": "2019-05-03T20:39:07.605Z",
              "surveyId": 112
              }
                          }
        
          schema:
           $ref: '#/definitions/SurveyResponseSuccess'
           
        '400':
          description: "the parametrs should not be empty"
          examples:
            aplication/json: {
              "message": "the parametrs should not be empty",
              "reqId": "2O3n-ffC_"
            } 
          schema:
            $ref: '#/definitions/SurveyResponseFail'
              
        '500':
          description: Unexpected error.    
          examples:
            aplication/json: {
              "message": "the parametrs should not be empty",
              "reqId": "gEA-5YfH1"
            }
          schema:
            $ref: '#/definitions/SurveyResponseError'
            
            
  ###############
  ### ANSWER  ###
  ###############
...
definitions:
  SurveyResponseSuccess:
    properties:
      response:
        type: object
        properties:
          survey:
            type: string
          surveyName:
            type: string
          createdAt: 
            type: string
          updatedAt:
            type: string
          surveyId:
            type: string 
...
  SurveyResponseFail:
    properties:
      response:
        type: string
      reqId:
        type: string
  SurveyResponseError:
    properties:
      response: 
        type: string
      reqId: 
        type: string
        
  # Answer
...